+++ frontend/src/components/layout/main/NewApplication.tsx
@@ -10,6 +10,7 @@ import Tab from "@mui/material/Tab";
 import Tabs from "@mui/material/Tabs";
 import Typography from "@mui/material/Typography";
 import React from "react";
+import { LoadingState } from "./LoadingState";
 
 import type { AlertColor } from '@mui/material/Alert';
 import { AxiosError } from 'axios';
@@ -175,7 +176,7 @@ export const NewApplication = () => {
             <Typography color="textSecondary" sx={{ mb: 4 }}>
                 Here you can start a new application for an authorisation process.
             </Typography>
-            {isQuestionnairesLoading ? <Typography>Loading questionnaires...</Typography> :
+            {isQuestionnairesLoading ? <LoadingState /> :
                 processGroups.length === 0 ? <EmptyStateComponent /> :
                     <>
                         {processGroups.map((group) => (